Move FSM runtime into dedicated module and feature flag
