+++ Cargo.toml
@@ -29,12 +29,12 @@ default = ["baselib", "pretty_print", "serde", "compiler", "program", "mika",
       "u8", "u16", "u32", "u64", "u128", 
       "i8", "i16", "i32", "i64", "i128", 
       "f32", "f64", "c64", "r64",
-      "statements_default", "subscript_default",
+      "statements_default", "subscript_default", "state_machines",
       "formatter", "mechfs", "serve", "run", "repl", "build", "whos", "async",
       "mech-core/default", "mech-interpreter/default", "mech-syntax/default",
       ]
 base = ["baselib", "pretty_print", "serde", "compiler", "program", "mika",
-      "statements_default", "subscript_default",
+      "statements_default", "subscript_default", "state_machines",
       "formatter", "mechfs", "serve", "run", "repl", "build","whos", "async",
       "mech-core/base", "mech-interpreter/base", "mech-syntax/base",
       ]
@@ -92,6 +92,7 @@ kind_define = ["kind_annotation", "statements", "mech-core/kind_define", "mech-i
 kind_annotation = ["functions", "mech-core/kind_annotation", "mech-interpreter/kind_annotation", "mech-syntax/kind_annotation"]
 formulas = ["mech-core/formulas", "mech-interpreter/formulas", "mech-syntax/formulas"]
 functions = ["symbol_table", "mech-core/functions", "mech-interpreter/functions", "mech-syntax/functions"]
+state_machines = ["functions", "mech-interpreter/state_machines"]
 symbol_table = ["mech-core/symbol_table", "mech-interpreter/symbol_table", "mech-syntax/symbol_table"]
 
 bool = ["mech-core/bool", "mech-interpreter/bool", "mech-syntax/bool"]
