+++ src/interpreter/src/expressions.rs
@@ -24,8 +24,8 @@ pub fn expression(expr: &Expression, env: Option<&Environment>, p: &Interpreter)
     Expression::SetComprehension(set_comp) => set_comprehension(set_comp, p),
     #[cfg(feature = "matrix_comprehensions")]
     Expression::MatrixComprehension(matrix_comp) => matrix_comprehension(matrix_comp, p),
-    #[cfg(feature = "functions")]
-    Expression::FsmPipe(fsm_pipe) => crate::functions::execute_fsm_pipe(fsm_pipe, env, p),
+    #[cfg(feature = "state_machines")]
+    Expression::FsmPipe(fsm_pipe) => crate::state_machines::execute_fsm_pipe(fsm_pipe, env, p),
     x => Err(MechError::new(
       FeatureNotEnabledError,
       None
