+++ src/interpreter/src/interpreter.rs
@@ -24,6 +24,7 @@ pub struct Interpreter {
   pub code: Vec<MechSourceCode>,
   pub out: Value,
   pub out_values: Ref<HashMap<u64, Value>>,
+  #[cfg(feature = "state_machines")]
   pub user_state_machines: Ref<HashMap<u64, FsmImplementation>>,
   pub sub_interpreters: Ref<HashMap<u64, Box<Interpreter>>>,
 }
@@ -44,6 +45,7 @@ impl Clone for Interpreter {
       code: self.code.clone(),
       out: self.out.clone(),
       out_values: self.out_values.clone(),
+      #[cfg(feature = "state_machines")]
       user_state_machines: self.user_state_machines.clone(),
       sub_interpreters: self.sub_interpreters.clone(),
     }
@@ -68,6 +70,7 @@ impl Interpreter {
       out: Value::Empty,
       sub_interpreters: Ref::new(HashMap::new()),
       out_values: Ref::new(HashMap::new()),
+      #[cfg(feature = "state_machines")]
       user_state_machines: Ref::new(HashMap::new()),
       code: Vec::new(),
       #[cfg(feature = "compiler")]
