+++ src/interpreter/src/lib.rs
@@ -80,6 +80,8 @@ pub mod interpreter;
 pub mod stdlib;
 #[cfg(feature = "functions")]
 pub mod functions;
+#[cfg(feature = "state_machines")]
+pub mod state_machines;
 pub mod statements;
 pub mod expressions;
 pub mod mechdown;
@@ -93,6 +95,8 @@ pub use crate::interpreter::*;
 pub use crate::structures::*;
 #[cfg(feature = "functions")]
 pub use crate::functions::*;
+#[cfg(feature = "state_machines")]
+pub use crate::state_machines::*;
 pub use crate::statements::*;
 pub use crate::expressions::*;
 pub use crate::mechdown::*;
